[Handler] Get file size, not using ffprobe, Fix path, drop unneeded code refs #1

+++ model.go
@@ -45,7 +45,7 @@ type Audio struct {
 	Id      uint64 `gorm:"column:id" json:"id,string"`
 	UserId  uint64 `gorm:"column:uid" json:"uid,string"`
 	Path    string `gorm:"column:path"`
-	Size    uint64 `gorm:"column:size"`
+	Size    int    `gorm:"column:size"`
 	Created string `gorm:"column:created" json:"created"`
 }
